WIP: pre-test checkpoint — extract SortRunnerService from label-view
Refs #3428.
