+++ frontend/src/app/components/label-view/label-view.component.ts
@@ -1,9 +1,7 @@
 import { AfterViewInit, ChangeDetectionStrategy, Component, DestroyRef, effect, ElementRef, inject, OnDestroy, OnInit, signal, untracked, viewChild } from '@angular/core';
 import { takeUntilDestroyed } from '@angular/core/rxjs-interop';
 
-import { HttpErrorResponse } from '@angular/common/http';
-import { Subscription, of, pairwise, throwError } from 'rxjs';
-import { catchError, filter, take, tap } from 'rxjs/operators';
+import { Subscription, pairwise } from 'rxjs';
 import { LeftPanelComponent } from '../left-panel/left-panel.component';
 import { CenterPanelComponent } from '../center-panel/center-panel.component';
 import { RightPanelComponent } from '../right-panel/right-panel.component';
@@ -19,16 +17,15 @@ import { NewThingFlowsService } from '../../services/new-thing-flows.service';
 import { ToastService } from '../../services/toast.service';
 import { SortingApiService } from '../../services/sorting-api.service';
 import { PairScopeService } from '../../services/pair-scope.service';
+import { SortRunnerService } from '../../services/sort-runner.service';
 import { adaptivePoll } from '../../services/adaptive-poll';
 import { DetectorsFindApiService } from '../../services/detectors-find-api.service';
 import { DetectorsRegistryApiService } from '../../services/detectors-registry-api.service';
-import { MediasApiService } from '../../services/medias-api.service';
 import { LabelSessionService } from '../../services/label-session.service';
 import { MediaStateService } from '../../services/media-state.service';
 import { VoteStateService } from '../../services/vote-state.service';
 import { LabelsetStateService } from '../../services/labelset-state.service';
 import { SortStateService, SortMode, SelectMode } from '../../services/sort-state.service';
-import { autoSelectNext as pickNextMedia } from '../../utils/auto-select-next';
 import { SettingsStateService } from '../../services/settings-state.service';
 import { AutopilotStateService } from '../../services/autopilot-state.service';
 import { EmbedderCapabilityService } from '../../services/embedder-capability.service';
@@ -37,7 +34,6 @@ import { DetectorRegistryEntry } from '../../generated/api-client/models/detecto
 import { ProgressModalComponent, ProgressMetric } from '../modals/progress-modal/progress-modal.component';
 import { ResortPromptModalComponent, ResortResult } from '../modals/resort-prompt-modal/resort-prompt-modal.component';
 import type { LabelingStatusResponse } from '../../generated/api-client/models/labeling-status-response';
-import type { LearnedSortResponse } from '../../generated/api-client/models/learned-sort-response';
 import { snapPanelWidthToGridColumns, iconSizeToGoalWidth } from '../../utils/grid-icon-size';
 import { PanelResizeDirective } from '../../directives/panel-resize.directive';
 import { LabelViewPanelStateService } from './label-view-panel-state.service';
@@ -57,7 +53,7 @@ import { buildMediaContextMenuItems } from './media-context-menu-items';
     MediaCropModalComponent,
     PanelResizeDirective
 ],
-  providers: [LabelViewPanelStateService, PairScopeService],
+  providers: [LabelViewPanelStateService, PairScopeService, SortRunnerService],
   templateUrl: './label-view.component.html',
   styleUrl: './label-view.component.scss',
 })
@@ -65,7 +61,6 @@ export class LabelViewComponent implements OnInit, AfterViewInit, OnDestroy {
   private sortingApi = inject(SortingApiService);
   private detectorsFindApi = inject(DetectorsFindApiService);
   private detectorsRegistryApi = inject(DetectorsRegistryApiService);
-  private mediasApi = inject(MediasApiService);
   private labelSession = inject(LabelSessionService);
   mediaState = inject(MediaStateService);
   voteState = inject(VoteStateService);
@@ -80,6 +75,9 @@ export class LabelViewComponent implements OnInit, AfterViewInit, OnDestroy {
   panelState = inject(LabelViewPanelStateService);
   /** Component-provided. Public: the header binds `pairScope.datasetName()`. */
   readonly pairScope = inject(PairScopeService);
+  /** Component-provided sort orchestration (#3428). Public: the handlers below
+   *  forward the template's events straight to it. */
+  private readonly sortRunner = inject(SortRunnerService);
 
   readonly layoutRef = viewChild.required<ElementRef<HTMLElement>>('layout');
   readonly centerPanel = viewChild(CenterPanelComponent);
@@ -98,8 +96,9 @@ export class LabelViewComponent implements OnInit, AfterViewInit, OnDestroy {
    *  Drives the center-pane "nothing left to label" message so the pane is
    *  never left blank with a stale item stuck in the metadata strip. */
   readonly autopilotExhausted = signal(false);
-  /** True while a windowed-sort "Load more" page fetch is in flight. */
-  readonly loadingMoreSort = signal(false);
+  /** True while a windowed-sort "Load more" page fetch is in flight. Aliased
+   *  from the runner that owns it, so the template binding is unchanged. */
+  readonly loadingMoreSort = this.sortRunner.loadingMoreSort;
   progressModalMetric: ProgressMetric | null = null;
 
   // SortStateService / VoteStateService are now signal-backed (their value
@@ -160,19 +159,6 @@ export class LabelViewComponent implements OnInit, AfterViewInit, OnDestroy {
   // than the pair scope, or `reloadForNewPair`'s teardown would kill the
   // request it just issued for the new pair.
   private statusPolling$: Subscription | null = null;
-  private learnedSortPending = false;
-  /** Active learned-sort job id while a training run is in flight. Set in
-   *  ``onLearnedSort`` once the backend returns a job id, cleared in
-   *  ``applyLearnedSortResult`` / the error/cancel paths. Used by the
-   *  Cancel button on the sort progress bar to target the right job. */
-  private currentLearnedSortJobId: string | null = null;
-  /** Consecutive *transient* learned-sort result-poll failures tolerated before
-   *  the run is declared failed. Roughly 10s–40s of unbroken failures at the
-   *  poll's 500ms–2000ms cadence: long enough to ride out a backend blip,
-   *  short enough that a genuinely unreachable server does not leave the panel
-   *  spinning on 'Training…'. Terminal statuses (404/500) end the run at once
-   *  and never consume this budget. */
-  private readonly POLL_ERROR_LIMIT = 20;
   /** Set by `reloadForNewPair` when the user was in `learned` sort mode at the
    *  time of a pair switch: a constructor effect watches the labelset counts and
    *  re-fires `onLearnedSort` once, after the reloaded votes make both classes
@@ -296,7 +282,7 @@ export class LabelViewComponent implements OnInit, AfterViewInit, OnDestroy {
         this.pendingSelectOnPairChange = false;
         // A re-rank that auto-selected on its own (or a click that beat us to
         // it) already owns the centre; never move the user off it.
-        if (this.mediaState.selectedId() === null) this.autoSelectNext();
+        if (this.mediaState.selectedId() === null) this.sortRunner.autoSelectNext();
       });
     });
 
@@ -313,7 +299,7 @@ export class LabelViewComponent implements OnInit, AfterViewInit, OnDestroy {
         if (!this.pendingRehydrateLearned) return;
         if (this.sortState.sortMode === 'learned' && this.voteState.learnedSortAvailable) {
           this.pendingRehydrateLearned = false;
-          this.onLearnedSort(false);
+          this.sortRunner.onLearnedSort(false);
         }
       });
     });
@@ -361,20 +347,20 @@ export class LabelViewComponent implements OnInit, AfterViewInit, OnDestroy {
           this.sortState.setSelectMode('top');
           if (curr.retrainMode) {
             this.sortState.setSortMode('learned');
-            this.onLearnedSort(false);
+            this.sortRunner.onLearnedSort(false);
           }
         }
         else if (curr.phase === 'bad') {
           this.sortState.setSelectMode('hard');
           if (curr.retrainMode) {
             this.sortState.setSortMode('learned');
-            this.onLearnedSort(false);
+            this.sortRunner.onLearnedSort(false);
           }
         }
         else if (curr.phase === 'hard') {
           this.sortState.setSelectMode('hard');
           this.sortState.setSortMode('learned');
-          this.onLearnedSort(false);
+          this.sortRunner.onLearnedSort(false);
         }
         else if (curr.phase === 'new') this.sortState.setSelectMode('new');
       });
@@ -402,9 +388,7 @@ export class LabelViewComponent implements OnInit, AfterViewInit, OnDestroy {
       // Train's scoring and learned-sort subscriptions carry no `finalize`, so
       // the busy flag, the job id and the progress feed they own are reset here
       // — after the supersede, so nothing can re-set them.
-      this.sortState.stopFindProgressTracking();
-      this.currentLearnedSortJobId = null;
-      this.sortState.setSortBusy(false);
+      this.sortRunner.quiesce();
       this.pendingRehydrateLearned = false;
       // Read by the medias effect when the reload below lands.
       this.pendingSnapOnLoad = true;
@@ -654,215 +638,29 @@ export class LabelViewComponent implements OnInit, AfterViewInit, OnDestroy {
   }
 
   // --- Sort handlers ---
+  //
+  // The sorts themselves live on `SortRunnerService` (component-provided, so it
+  // can use the same pair scope this view does). What is left here is the
+  // template's event surface: one line each, no logic.
 
   onSortModeChange(mode: SortMode): void {
-    this.sortState.setSortMode(mode);
-    this.autoSelectNext();
+    this.sortRunner.onSortModeChange(mode);
   }
 
-  /**
-   * Install a (possibly windowed) sort response into the sort state. Handles
-   * both `similarity` (text/example sort) and `score` (learned/detector sort)
-   * result rows, and carries the window metadata (`total` / `has_more_below` /
-   * `sort_token`) so the media-list can page deeper. Below the backend's window
-   * threshold the whole ranking arrives and `has_more_below` is false —
-   * behaviour is identical to the pre-windowing full-list path.
-   */
-  private applySortWindow(response: {
-    results?: Array<Record<string, unknown>>;
-    threshold?: number;
-    acq_threshold?: number | null;
-    total?: number;
-    above_threshold?: number;
-    has_more_below?: boolean;
-    sort_token?: string;
-  }): void {
-    const threshold = response.threshold ?? 0;
-    const items = (response.results ?? []).map((r) => ({
-      id: r['id'] as number,
-      score: (r['score'] ?? r['similarity'] ?? 0) as number,
-      bestRegion: r['best_region'] as number[] | undefined,
-    }));
-    this.sortState.setSortWindow({
-      items,
-      threshold,
-      acqThreshold: response.acq_threshold ?? null,
-      total: response.total ?? items.length,
-      hasMore: response.has_more_below ?? false,
-      token: response.sort_token ?? null,
-      aboveThreshold: response.above_threshold ?? items.filter((i) => i.score >= threshold).length,
-    });
-  }
-
-  /**
-   * Page in the next window of a windowed ranking (the media-list "Load more"
-   * trigger). Fetches from the sort token at the current loaded offset and
-   * appends. A failed/expired token just stops paging (the user can re-sort).
-   */
   onLoadMore(): void {
-    const token = this.sortState.sortToken;
-    if (!token || !this.sortState.sortHasMore || this.loadingMoreSort()) return;
-    this.loadingMoreSort.set(true);
-    const offset = this.sortState.sortOrder?.length ?? 0;
-    this.sortingApi
-      .getSortPage(token, offset, 200)
-      .pipe(this.pairScope.scoped())
-      .subscribe({
-        next: (page) => {
-          const items = (page.results ?? []).map((r) => ({
-            id: r['id'] as number,
-            score: (r['score'] ?? r['similarity'] ?? 0) as number,
-            bestRegion: r['best_region'] as number[] | undefined,
-          }));
-          this.sortState.appendSortItems(items, page.has_more);
-          this.loadingMoreSort.set(false);
-        },
-        error: () => this.loadingMoreSort.set(false),
-      });
+    this.sortRunner.onLoadMore();
   }
 
   onTextSort(text: string): void {
-    this.sortState.setTextQuery(text);
-    this.sortState.setSortBusy(true);
-    this.sortState.setSortStatus('Sorting…');
-    this.sortingApi.sort({ text }).pipe(this.pairScope.scoped()).subscribe({
-      next: (response) => {
-        this.applySortWindow(response);
-        this.sortState.setSortBusy(false);
-        this.sortState.setSortStatus('');
-        this.autoSelectNext();
-      },
-      error: () => {
-        this.sortState.setSortBusy(false);
-        this.sortState.setSortStatus('Sort failed');
-      },
-    });
+    this.sortRunner.onTextSort(text);
   }
 
   onLearnedSort(autoSelect = true): void {
-    if (!this.voteState.learnedSortAvailable) return;
-    this.sortState.setSortBusy(true);
-    this.sortState.setSortStatus('Training…');
-    this.sortingApi.learnedSort().pipe(this.pairScope.scoped()).subscribe({
-      next: (response) => {
-        if (response.status === 'done') {
-          this.applyLearnedSortResult(response, autoSelect);
-        } else if (response.status === 'running') {
-          this.currentLearnedSortJobId = response.job_id;
-          this.pollLearnedSortJob(response.job_id, autoSelect);
-        } else {
-          this.sortState.setSortBusy(false);
-          this.sortState.setSortStatus(response.error || 'Training failed');
-        }
-      },
-      error: () => {
-        this.sortState.setSortBusy(false);
-        this.sortState.setSortStatus('Training failed');
-      },
-    });
+    this.sortRunner.onLearnedSort(autoSelect);
   }
 
-  /**
-   * Poll a running learned-sort job until it settles.
-   *
-   * Uses {@link adaptivePoll}, not the `timer(200, 500)` + `switchMap` pattern
-   * this once had: `switchMap` aborted the in-flight result GET on every tick,
-   * so a backend that needed longer than the interval to answer — exactly the
-   * situation while an MLP training job is hogging the process — had *every*
-   * read cancelled, never saw a non-running status, and left the panel stuck
-   * on 'Training…' with `sortBusy` true forever. That is the pathology
-   * documented in `adaptive-poll.ts` (issue #2572) that the labeling-status
-   * poll above was already migrated off; this poll was left behind.
-   *
-   * Poll failures are no longer fatal either. The result endpoint reports two
-   * genuine terminal states by HTTP status code — 404 (job evicted or unknown)
-   * and 500 (the job itself errored) — so those still end the run, but any
-   * other failure (a network blip, a proxy 502/503) is transient and costs
-   * only that tick, until {@link POLL_ERROR_LIMIT} consecutive failures say
-   * the backend is really gone. Previously a single transient error tore the
-   * poll down and reported 'Training failed' for a job still running
-   * server-side.
-   */
-  private pollLearnedSortJob(jobId: string, autoSelect: boolean): void {
-    let consecutiveErrors = 0;
-    const settledWith = (error: string): LearnedSortResponse => ({
-      job_id: jobId,
-      status: 'error',
-      error,
-    });
-
-    adaptivePoll<LearnedSortResponse>(
-      () =>
-        this.sortingApi.getLearnedSortResult(jobId).pipe(
-          tap(() => (consecutiveErrors = 0)),
-          catchError((err: unknown) => {
-            const status = err instanceof HttpErrorResponse ? err.status : 0;
-            if (status === 404) return of(settledWith('Training job expired'));
-            if (status === 500) return of(settledWith('Training failed'));
-            consecutiveErrors += 1;
-            if (consecutiveErrors >= this.POLL_ERROR_LIMIT) {
-              return of(settledWith('Training failed'));
-            }
-            // Re-throw so adaptivePoll absorbs it: this tick is skipped and the
-            // next one scheduled as usual, rather than the poll tearing down.
-            return throwError(() => err);
-          }),
-        ),
-      { fastMs: 500, slowMs: 2000 },
-    )
-      .pipe(
-        // Pair-scoped: a training job can outlive the pair it was started for,
-        // and its result must not be applied to whatever pair is active when it
-        // finally settles (see `PairScopeService`).
-        this.pairScope.scoped(),
-        filter((res) => res.status !== 'running'),
-        take(1),
-      )
-      // No `error` handler: adaptivePoll never errors — a request failure is
-      // either absorbed above or converted into a terminal `error` status.
-      .subscribe((res) => {
-        if (res.status === 'done') {
-          this.applyLearnedSortResult(res, autoSelect);
-        } else if (res.status === 'cancelled') {
-          this.currentLearnedSortJobId = null;
-          this.sortState.setSortBusy(false);
-          this.sortState.setSortStatus('Cancelled');
-        } else {
-          this.currentLearnedSortJobId = null;
-          this.sortState.setSortBusy(false);
-          this.sortState.setSortStatus(res.error || 'Training failed');
-        }
-      });
-  }
-
-  private applyLearnedSortResult(response: LearnedSortResponse, autoSelect: boolean): void {
-    this.applySortWindow(response);
-    this.currentLearnedSortJobId = null;
-    this.sortState.setSortBusy(false);
-    this.sortState.setSortStatus('');
-    if (autoSelect) {
-      this.autoSelectNext();
-    }
-  }
-
-  /** Cancel whatever sort run is currently in flight.
-   *
-   *  - Learned sort: targets the active ``AsyncJob`` by id.
-   *  - Load-sort (find-label): trips the shared ``find_progress`` cancel
-   *    flag, which the scoring loop polls.
-   *  - Text / example sort: no cancellation endpoint; those calls run
-   *    synchronously and complete before the user can usefully cancel.
-   */
   onSortCancel(): void {
-    if (this.currentLearnedSortJobId) {
-      const jobId = this.currentLearnedSortJobId;
-      this.currentLearnedSortJobId = null;
-      this.sortingApi.cancelLearnedSort(jobId).pipe(takeUntilDestroyed(this.destroyRef)).subscribe();
-      return;
-    }
-    if (this.sortState.sortMode === 'load') {
-      this.detectorsFindApi.cancelFind().pipe(takeUntilDestroyed(this.destroyRef)).subscribe();
-    }
+    this.sortRunner.onSortCancel();
   }
 
   onLoadSort(): void {
@@ -870,101 +668,23 @@ export class LabelViewComponent implements OnInit, AfterViewInit, OnDestroy {
   }
 
   onModelSelected(modelId: string): void {
-    if (!modelId) return;
-    this.sortState.setSortMode('load');
-    this.sortState.setSortBusy(true);
-    this.sortState.setSortStatus('Scoring with detector…');
-    this.sortState.setSortProgress(0, 0);
-
-    this.sortState.startFindProgressTracking();
-
-    // Pair-scoped: scoring runs for minutes on a large dataset, so a pair switch
-    // mid-run must kill this before it ranks the new pair with old scores.
-    this.detectorsFindApi.findLabel({ detector_id: modelId }).pipe(this.pairScope.scoped()).subscribe({
-      next: (raw) => {
-        const response = raw as {
-          results: { id: number; score: number; best_region?: number[] }[];
-          threshold: number;
-          detector_name?: string;
-        };
-        this.sortState.stopFindProgressTracking();
-        this.applySortWindow(response);
-        this.sortState.setLoadSortLabel(response.detector_name || 'Detector');
-        this.sortState.setSortBusy(false);
-        this.sortState.setSortStatus('');
-        this.sortState.setSortProgress(0, 0);
-        this.autoSelectNext();
-      },
-      error: () => {
-        this.sortState.stopFindProgressTracking();
-        this.sortState.setSortBusy(false);
-        this.sortState.setSortStatus('Detector sort failed');
-        this.sortState.setSortProgress(0, 0);
-      },
-    });
+    this.sortRunner.onModelSelected(modelId);
   }
 
   onExampleSortStarted(data: unknown): void {
-    const response = data as {
-      results: { id: number; similarity: number; best_region?: number[] }[];
-      threshold: number;
-    };
-    this.sortState.setSortMode('load');
-    this.applySortWindow(response);
-    this.sortState.setLoadSortLabel('Example media');
-    this.sortState.setSortBusy(false);
-    this.sortState.setSortStatus('');
-    this.autoSelectNext();
+    this.sortRunner.onExampleSortStarted(data);
   }
 
   // --- Select mode ---
 
   onSelectModeChange(mode: SelectMode): void {
-    this.sortState.setSelectMode(mode);
-    this.autoSelectNext();
-  }
-
-  private fetchDiversityNext(): void {
-    const sortOrder = this.sortState.sortOrder;
-    const scores = sortOrder
-      ? Object.fromEntries(sortOrder.map((s) => [String(s.id), s.score]))
-      : undefined;
-    this.sortingApi
-      // The New pick reads the threshold as a sampling position too (it steers
-      // the atlas probe by a node's median score), so it takes the acquisition
-      // cut alongside the Hard pick.
-      .getCoverageAtlasNext(scores, this.sortState.acqThreshold ?? undefined)
-      .pipe(this.pairScope.scoped())
-      .subscribe({
-        next: (response) => {
-          if (response.id !== null) {
-            this.mediaState.selectMedia(response.id);
-          }
-          if (typeof response.coverage_level === 'number') {
-            this.autopilotStateService.updateDiversityLevel(response.coverage_level);
-          }
-        },
-      });
+    this.sortRunner.onSelectModeChange(mode);
   }
 
   // --- Inclusion ---
 
   onInclusionChange(value: number): void {
-    this.sortState.setInclusion(value);
-    this.sortingApi.setInclusion(value).pipe(this.pairScope.scoped()).subscribe();
-    this.autoSelectNext();
-    if (this.sortState.sortMode === 'learned' && this.voteState.learnedSortAvailable) {
-      this.scheduleLearnedSort(false);
-    }
-  }
-
-  private scheduleLearnedSort(autoSelect = true): void {
-    if (this.learnedSortPending) return;
-    this.learnedSortPending = true;
-    setTimeout(() => {
-      this.learnedSortPending = false;
-      this.onLearnedSort(autoSelect);
-    }, 300);
+    this.sortRunner.onInclusionChange(value);
   }
 
   // --- Media selection ---
@@ -1020,26 +740,7 @@ export class LabelViewComponent implements OnInit, AfterViewInit, OnDestroy {
   }
 
   private runExampleSortById(mediaId: number, cropParams?: Record<string, unknown>): void {
-    this.sortState.setSortBusy(true);
-    this.sortState.setSortStatus('Sorting by example…');
-    this.sortingApi
-      .exampleSortById({ media_id: mediaId, crop_params: cropParams })
-      .pipe(this.pairScope.scoped())
-      .subscribe({
-        next: (response) => {
-          this.sortState.setSortMode('load');
-          this.applySortWindow(response);
-          this.sortState.setLoadSortLabel(this.mediaDisplayName(mediaId));
-          this.sortState.setSortBusy(false);
-          this.sortState.setSortStatus('');
-          this.autoSelectNext();
-        },
-        error: (err) => {
-          this.sortState.setSortBusy(false);
-          this.sortState.setSortStatus('Example sort failed');
-          this.toast.error({ message: err?.error?.message || 'Example sort failed' });
-        },
-      });
+    this.sortRunner.runExampleSortById(mediaId, this.mediaDisplayName(mediaId), cropParams);
   }
 
   private openSeedNewDetector(mediaId: number, cropParams?: Record<string, unknown>): void {
@@ -1141,9 +842,9 @@ export class LabelViewComponent implements OnInit, AfterViewInit, OnDestroy {
     // soon as the server has it, instead of waiting for the next poll tick.
     // No-op when no detector is being trained (refresh() bails on a null model).
     this.labelsetState.refresh();
-    this.autoSelectNext(event.id);
+    this.sortRunner.autoSelectNext(event.id);
     if (this.sortState.sortMode === 'learned' && this.voteState.learnedSortAvailable) {
-      this.scheduleLearnedSort(false);
+      this.sortRunner.scheduleLearnedSort(false);
     }
     this.checkResortPrompt();
   }
@@ -1238,7 +939,7 @@ export class LabelViewComponent implements OnInit, AfterViewInit, OnDestroy {
         }
       } else {
         // No sort query configured; try to select from existing sort results.
-        this.autoSelectNext();
+        this.sortRunner.autoSelectNext();
       }
     }
   }
@@ -1251,31 +952,11 @@ export class LabelViewComponent implements OnInit, AfterViewInit, OnDestroy {
     // left-panel disables the Autopilot tab (see `autopilotDisabled`) and the
     // user labels manually until Learn sort re-enables Autopilot.
     if (!this.textSupported) return;
-    this.onTextSort(textQuery);
+    this.sortRunner.onTextSort(textQuery);
   }
 
   private triggerAutopilotMediaSort(): void {
-    // Every media example seeds the sort: plural examples rank the haystack
-    // against the centroid of their embeddings, so the "good" phase surfaces
-    // items resembling what the examples have in common.
-    const filenames = this.labelSession.mediaExampleFilenames;
-    if (filenames.length > 0) {
-      this.sortState.setSortBusy(true);
-      this.sortState.setSortStatus(filenames.length > 1 ? 'Sorting by examples…' : 'Sorting by example…');
-      this.sortingApi.exampleSortServer({ filenames }).pipe(this.pairScope.scoped()).subscribe({
-        next: (response) => {
-          this.applySortWindow(response);
-          this.sortState.setSortBusy(false);
-          this.sortState.setSortStatus('');
-          this.sortState.setSortMode('load');
-          this.autoSelectNext();
-        },
-        error: () => {
-          this.sortState.setSortBusy(false);
-          this.sortState.setSortStatus('Example sort failed');
-        },
-      });
-    }
+    this.sortRunner.exampleSortByFilenames(this.labelSession.mediaExampleFilenames);
   }
 
   // --- Re-sort prompt ---
@@ -1347,7 +1028,7 @@ export class LabelViewComponent implements OnInit, AfterViewInit, OnDestroy {
   }
 
   onAutopilotRefocus(): void {
-    this.autoSelectNext();
+    this.sortRunner.autoSelectNext();
   }
 
   onAutopilotToggleCollapse(): void {
@@ -1427,30 +1108,5 @@ export class LabelViewComponent implements OnInit, AfterViewInit, OnDestroy {
     }
   }
 
-  // --- Helpers ---
 
-  /**
-   * Advance to the next media the current Sort + Select says to show.
-   *
-   * The rule itself lives in {@link autoSelectNext} (`utils/auto-select-next`)
-   * as a pure function so it can be unit-tested and so the eval harness's copy
-   * of it can be digest-pinned; this method is the side-effecting half —
-   * applying the selection, or firing the coverage-atlas probe the `new` mode
-   * asks for.
-   */
-  private autoSelectNext(excludeId?: number): void {
-    const pick = pickNextMedia({
-      sortOrder: this.sortState.sortOrder,
-      selectMode: this.sortState.selectMode,
-      acqThreshold: this.sortState.acqThreshold,
-      goodVotes: this.voteState.goodVotes,
-      badVotes: this.voteState.badVotes,
-      excludeId,
-    });
-    if (pick.kind === 'media') {
-      this.mediaState.selectMedia(pick.id);
-    } else if (pick.kind === 'diversity') {
-      this.fetchDiversityNext();
-    }
-  }
 }
